+++ engine/src/fusion/graphics/renderpass/framebuffers.hpp
@@ -17,10 +17,10 @@ namespace fe {
         ~Framebuffers();
         NONCOPYABLE(Framebuffers);
 
-        const Image2d* getAttachment(uint32_t index) const { return imageAttachments[index].get(); }
+        const Image2d* getAttachment(size_t index) const { return imageAttachments[index].get(); }
 
-        const VkFramebuffer& getFramebuffer(uint32_t index) const { return framebuffers[index < framebuffers.size() ? index : 0]; }
-        const VkFramebuffer& operator[](uint32_t index) const { return framebuffers[index < framebuffers.size() ? index : 0]; }
+        const VkFramebuffer& getFramebuffer(size_t index) const { return framebuffers[index < framebuffers.size() ? index : 0]; }
+        const VkFramebuffer& operator[](size_t index) const { return framebuffers[index < framebuffers.size() ? index : 0]; }
 
         const std::vector<std::unique_ptr<Image2d>>& getImageAttachments() const { return imageAttachments; }
         const std::vector<VkFramebuffer>& getFramebuffers() const { return framebuffers; }
